move CustomIcons components out of Geral subfolder
