+++ website/src/Components/CustomIcons/OcrIcon.js
@@ -5,10 +5,10 @@ export default function OcrIcon(props) {
     return (
         <SvgIcon
             {...props}
-            version="1.0" 
+            version="1.0"
             xmlns="http://www.w3.org/2000/svg"
-            width="512.000000pt" 
-            height="512.000000pt" 
+            width="512.000000pt"
+            height="512.000000pt"
             viewBox="0 0 512.000000 512.000000"
             preserveAspectRatio="xMidYMid meet"
         >
